Added additional Scaffolding
The preferences now work so that it can tell if a current configuration
is set and display the appropriate page. It will need to draw the data
to display on the configuration page from the scheduler class.
Added a title to each page.
Added a delete button to configuration page.